requirements.txt
# Python Dependencies - Simple Version (NO API/Flask)
# Install with: pip install -r requirements.txt

psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
